Remove use of value type.
