+++ packages/datajoint-core/src/results/mod.rs
@@ -1,7 +1,5 @@
 mod table_column;
 mod table_row;
-mod value;
 
 pub use table_column::{ColumnIndex, TableColumn, TableColumnRef};
 pub use table_row::TableRow;
-pub use value::{Value, ValueDecodable, ValueRef};
